Setup welcome screen skeleton
