+++ lib/src/app.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+import 'package:red_tv_youtube/src/screens/welcome.dart';
+
+class MyApp extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+        visualDensity: VisualDensity.adaptivePlatformDensity,
+      ),
+      home: WelcomeScreen(title: 'Flutter Demo Home Page'),
+    );
+  }
+}
